Moved CTags parser out of legacy
