+++ arduino/builder/preprocessor/ctags/testdata/TestCTagsParserFunctionPointersNoIndirect.txt
